split up game files
